+++ src/connectedk8s/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+1.10.0
+++++++
+* Added support for arc gateway feature.
+* Bug fix for Workload Identity poll or agent state
+* Bug fix for connectedk8s update with workload identity enabled
+
 1.9.0
 ++++++
 * Upgrading to v0.2.2 of ArcK8sDiagnoser